+++ doc/ref/examples/ex_objc_table_typed_intro.m
@@ -2,42 +2,53 @@
 #import <Tightdb/Tightdb.h>
 #import "people.h"
 
-/* PeopleTable is declared in people.h as
+/*
+ The classes People, PeopleQuery, PeopleView, and PeopleRow are declared 
+ (interfaces are generated) in people.h as
 
- TIGHTDB_TABLE_3(PeopleTable,
-                 Name,  String,
-                 Age,   Int,
-                 Hired, Bool);
+ TIGHTDB_TABLE_DEF_3(People,
+                     Name,  String,
+                     Age,   Int,
+                     Hired, Bool)
+ 
+ and in people.m you must have
+ 
+ TIGHTDB_TABLE_IMPL_3(People,
+                      Name, String,
+                      Age,  Int,
+                      Hired, Bool)
+ 
+in order to generate the implementation of the classes.
 */
 
 void ex_objc_table_typed_intro()
 {
-    /* Create a new table of the type defined above. */
-    PeopleTable *table = [[PeopleTable alloc] init];
+    // Create a new table of the type defined above
+    People *table = [[People alloc] init];
     
-    /* Append three rows. */
+    // Append three rows
     [table addRow:@{@"Name":@"Brian",  @"Age":@10,  @"Hired":@NO}];
     [table addRow:@{@"Name":@"Sofie",  @"Age":@40,  @"Hired":@YES}];
     [table addRow:@{@"Name":@"Jesper", @"Age":@200, @"Hired":@NO}];
     NSLog(@"The size of the table is now %zd", table.rowCount);
 
-    for (PeopleTableRow *ite in table) {
-        NSLog(@"Name: %@ Age: %lli", ite.Name, ite.Age);
+    for (PeopleRow *row in table) {
+        NSLog(@"Name: %@ Age: %lli", row.Name, row.Age);
     }
     
     NSLog(@"Insert a new row");
     [table insertRow:@{@"Name":@"Sam", @"Age":@30, @"Hired":@YES}
              atIndex:1];
     
-    for (PeopleTableRow *row in table) {
+    for (PeopleRow *row in table) {
         NSLog(@"Name: %@ Age: %lli", row.Name, row.Age);
     }
     
-    TDBRow *row2 = [table rowAtIndex:table.rowCount-1];
+    PeopleRow *row2 = [table rowAtLastIndex];
     if (row2 != nil)
         NSLog(@"Last row");
     
-    TDBRow *row3 = [table rowAtIndex:table.rowCount];
+    PeopleRow *row3 = [table rowAtIndex:42];
     if (row3 == nil)
         NSLog(@"Index out of range.");
 }