+++ doc/ref/examples/examples.h
@@ -7,3 +7,5 @@ void ex_objc_table_typed_intro();
 void ex_objc_table_typed_intro_with_many_comments();
 void ex_objc_tableview_dynamic_intro();
 void ex_objc_tableview_typed_intro();
+void ex_objc_intro();
+void ex_objc_table_dyn_table_sizes();